Prüfen, wie oft die Beans gecallt wurde

+++ src/main/java/org/spring/proxy/test/ThreadScopedBean.java
@@ -12,8 +12,8 @@ public class ThreadScopedBean {
 
     // Getter / Setter
 
-    public AtomicInteger getCount() {
-        return count;
+    public int getCount() {
+        return count.intValue();
     }
 
 }
